Refactored exceptions in com.softserve.osbb.util.resources.exceptions
