--- web/src/main/java/com/softserve/osbb/util/EntityNotFoundException.java
@@ -1,26 +1,0 @@
-package com.softserve.osbb.util;
-
-/**
- * Created by Anastasiia Fedorak on 13.07.2016.
- */
-public class EntityNotFoundException extends Exception {
-    public EntityNotFoundException() {
-    }
-
-    public EntityNotFoundException(String message) {
-
-        super(message);
-    }
-
-    public EntityNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public EntityNotFoundException(Throwable cause) {
-        super(cause);
-    }
-
-    public EntityNotFoundException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
